+++ c++14-opengl/shader/solar_rk_end.cs
@@ -4,11 +4,11 @@ layout(local_size_x = 1, local_size_y = 1, local_size_z = 1)in;
 // 質点
 struct Point
 {
-  float mass; // 質量
-  vec3 position; // 位置
-  vec3 velocity; // 速度
-  vec3 position_temp; // 位置(最終結果)
-  vec3 velocity_temp; // 速度(最終結果)
+  double mass; // 質量
+  dvec3 position; // 位置
+  dvec3 velocity; // 速度
+  dvec3 position_temp; // 位置(最終結果)
+  dvec3 velocity_temp; // 速度(最終結果)
 };
 
 // 質点群(仮値A)
@@ -26,31 +26,31 @@ layout(std430, binding = 1) buffer WritePoints
 layout (std140) uniform PhysicParams
 {
   uint point_num; // 質点数
-  float dt; // タイムステップ
-  float g; // 重力加速度
-  float r_threshold; // 引力が発生する距離の閾値
+  double dt; // タイムステップ
+  double g; // 重力加速度
+  double r_threshold; // 引力が発生する距離の閾値
 };
 
-uniform float x_dt2; // 更新時にdtに掛ける係数(最終結果計算用)
+uniform double x_dt2; // 更新時にdtに掛ける係数(最終結果計算用)
 
-vec3 calc_accel(uint i)
+dvec3 calc_accel(uint i)
 {
-  vec2 a = vec2(0.0);
+  dvec2 a = dvec2(0.0);
 
   // 自分以外の全質点に対して相互距離が一定以上なら
   // 相互距離の一乗に反比例する万有引力(/自分の質量)を計算して加算
   for (uint j = 0; j < i; ++j) {
-    vec2 dpos = current_points[j].position.xy - current_points[i].position.xy;
-    float r = length(dpos);
+    dvec2 dpos = current_points[j].position.xy - current_points[i].position.xy;
+    double r = length(dpos);
     a += step(r_threshold, r) * normalize(dpos) * g * current_points[j].mass / r;
   }
   for (uint j = i + 1; j < point_num; ++j) {
-    vec2 dpos = current_points[j].position.xy - current_points[i].position.xy;
-    float r = length(dpos);
+    dvec2 dpos = current_points[j].position.xy - current_points[i].position.xy;
+    double r = length(dpos);
     a += step(r_threshold, r) * normalize(dpos) * g * current_points[j].mass / r;
   }
 
-  return vec3(a, 0.0);
+  return dvec3(a, 0.0);
 }
 
 // 仮値Aから加速度(力)を計算して最終的な次フレームの速度/位置を作成
@@ -60,10 +60,10 @@ void main()
   // 計算対象のindex
   const uint i = gl_WorkGroupID.x;
 
-  vec3 a = calc_accel(i);
+  dvec3 a = calc_accel(i);
 
   // 最終結果の速度/位置を更新
-  vec3 delta = x_dt2 * dt * a;
+  dvec3 delta = x_dt2 * dt * a;
   next_points[i].velocity = current_points[i].velocity_temp + delta;
   next_points[i].velocity_temp = current_points[i].velocity_temp + delta;
