+++ c++14-opengl/shader/solar_rk_init.cs
@@ -4,11 +4,11 @@ layout(local_size_x = 1, local_size_y = 1, local_size_z = 1)in;
 // 質点
 struct Point
 {
-  float mass; // 質量
-  vec3 position; // 位置
-  vec3 velocity; // 速度
-  vec3 position_temp; // 位置(最終結果)
-  vec3 velocity_temp; // 速度(最終結果)
+  double mass; // 質量
+  dvec3 position; // 位置
+  dvec3 velocity; // 速度
+  dvec3 position_temp; // 位置(最終結果)
+  dvec3 velocity_temp; // 速度(最終結果)
 };
 
 // 質点群(元値)
@@ -26,9 +26,9 @@ layout(std430, binding = 1) buffer WritePoints
 layout (std140) uniform PhysicParams
 {
   uint point_num; // 質点数
-  float dt; // タイムステップ
-  float g; // 重力加速度
-  float r_threshold; // 引力が発生する距離の閾値
+  double dt; // タイムステップ
+  double g; // 重力加速度
+  double r_threshold; // 引力が発生する距離の閾値
 };
 
 void main()
